+++ mlp/src/main/java/mlp/layers/Layer.java
@@ -6,5 +6,5 @@ import org.jblas.DoubleMatrix;
 
 public interface Layer {
 	public DoubleMatrix forward(DoubleMatrix v) throws IllegalAccessException, IllegalArgumentException, InvocationTargetException;
-	public DoubleMatrix back(DoubleMatrix output_error, float learning_rate) throws IllegalAccessException, IllegalArgumentException, InvocationTargetException; 
+	public DoubleMatrix back(DoubleMatrix output_error, double learning_rate) throws IllegalAccessException, IllegalArgumentException, InvocationTargetException; 
 }
